pyproject.toml
[project]
name = "iqss-feasibility"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "blinker>=1.9.0",
    "fastapi>=0.115.12",
    "mypy>=1.16.1",
    "openpyxl>=3.1.5",
    "oracledb>=3.1.0",
    "pandas>=2.2.3",
    "plotly>=6.0.1",
    "pyodbc>=5.2.0",
    "python-dotenv>=1.1.0",
    "ruff>=0.11.9",
    "streamlit>=1.45.0",
    "uvicorn>=0.34.2",
    "pyjwt (>=2.10.1,<3.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "bcrypt (>=4.3.0,<5.0.0)",
    "streamlit-javascript (>=0.1.5,<0.2.0)",
]

[tool.mypy]
python_version = "3.12"
# Strict guidelines taken from https://github.com/pytorch/pytorch/blob/master/mypy-strict.ini
cache_dir = ".mypy_cache/strict"
strict_optional = true
hide_error_codes = false
show_column_numbers = true
warn_no_return = true
disallow_any_unimported = true
# Across versions of mypy, the flags toggled by --strict vary.  To ensure
# we have reproducible type check, we instead manually specify the flags
warn_unused_configs = true
disallow_any_generics = true
disallow_subclassing_any = true
disallow_untyped_calls = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
implicit_optional = false
warn_redundant_casts = true
warn_return_any = true
implicit_reexport = false
strict_equality = true
[tool.pydantic-mypy]
init_forbid_extra = true
init_typed = true
warn_required_dynamic_aliases = true
warn_untyped_fields = true
[tool.pytest.ini_options]
log_cli = true
log_cli_level = 10
testpaths = ["tests"]
addopts = "-p no:cacheprovider --cache-clear"
env = [
    "ENV_DESC=dev",
]
[tool.coverage.run]
source = ["tests"]
[tool.coverage.report]
skip_empty = true
sort = "Cover"

[tool.ruff]
line-length = 120
target-version = 'py312'

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4",  # flake8-comprehensions
    "PTH",  # flake8-use-pathlib
    "UP",  # pyupgrade
    "FAST",  # fastapi
    "AIR",  # airflow
    "C90",  # mccabe complexity
    "ERA",  # eradicate
]
ignore = [
    "PTH123"  # use Path.open() instead of builtin open()
]
